split specs
